+++ stage3/xfce/stage3-xfce.sh
@@ -553,7 +553,7 @@ systemctl enable lightdm
 cd ..
 rm -rf lightdm-gtk-greeter-2.0.8
 # Firefox.
-tar --no-same-owner -xf firefox-105.0.tar.bz2 -C /usr/lib
+tar --no-same-owner -xf firefox-105.0.1.tar.bz2 -C /usr/lib
 mkdir -p /usr/lib/firefox/distribution
 cat > /usr/lib/firefox/distribution/policies.json << END
 {